requirements.txt
loguru==0.7.2
python-dotenv==1.0.1
pydantic-settings==2.6.0
wandb==0.18.5
numpy==2.0.2
pandas==2.2.3
websockets==15.0.1
requests==2.31.0
psutil==7.0.0
asyncio==3.4.3
pymongo==4.11.2
pyopenssl==24.2.1
paramiko==3.5.0
asyncssh==2.19.0
joblib==1.4.2
aiohttp
scikit-learn==1.6.0
ansible-vault==2.1.0
bittensor==9.4.0
bittensor-cli==9.4.0
azure-identity==1.23.0
azure-mgmt-compute==34.1.0
azure-mgmt-network==29.0.0